+++ godot-macros/src/godot_cfg.rs
@@ -1,12 +1,19 @@
+use crate::godot_cfg::all::*;
+use crate::godot_cfg::any::*;
+use crate::godot_cfg::list::*;
+use crate::godot_cfg::not::*;
+use crate::godot_cfg::option::*;
 use proc_macro2::{Group, Ident, TokenTree};
 use std::error::Error;
 use std::fmt::{Display, Formatter};
-use std::ops::Deref;
-use crate::godot_cfg::option::{GodotConfigurationOption, GodotConfigurationOptionError};
 
+mod all;
+mod any;
+mod list;
+mod not;
 mod option;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum GodotConditionalCompilationError {
     UnableToParse(String),
     VenialError(venial::Error),
@@ -18,13 +25,19 @@ impl From<GodotConfigurationOptionError> for GodotConditionalCompilationError {
     }
 }
 
+impl From<GodotConfigurationPredicateListError> for GodotConditionalCompilationError {
+    fn from(value: GodotConfigurationPredicateListError) -> Self {
+        Self::VenialError(value.into())
+    }
+}
+
 impl Display for GodotConditionalCompilationError {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
             Self::UnableToParse(e) => {
                 write!(f, "Unable to parse '{}'", e)
             }
-            Self::VenialError(e) => write!(f, "{}", e)
+            Self::VenialError(e) => write!(f, "{}", e),
         }
     }
 }
@@ -55,11 +68,35 @@ impl GodotConditionalCompilation for GodotConfigurationPredicate {
     }
 }
 
+impl From<GodotConfigurationOption> for GodotConfigurationPredicate {
+    fn from(option: GodotConfigurationOption) -> Self {
+        Self::Option(option)
+    }
+}
+
+impl From<GodotConfigurationAll> for GodotConfigurationPredicate {
+    fn from(all: GodotConfigurationAll) -> Self {
+        Self::All(Box::new(all))
+    }
+}
+
+impl From<GodotConfigurationAny> for GodotConfigurationPredicate {
+    fn from(any: GodotConfigurationAny) -> Self {
+        Self::Any(Box::new(any))
+    }
+}
+
+impl From<GodotConfigurationNot> for GodotConfigurationPredicate {
+    fn from(not: GodotConfigurationNot) -> Self {
+        Self::Not(Box::new(not))
+    }
+}
+
 impl TryFrom<&Ident> for GodotConfigurationPredicate {
     type Error = GodotConditionalCompilationError;
 
     fn try_from(ident: &Ident) -> Result<Self, Self::Error> {
-        Ok(Self::Option(GodotConfigurationOption::try_from(ident)?))
+        Ok(GodotConfigurationOption::try_from(ident)?.into())
     }
 }
 
@@ -79,9 +116,9 @@ impl TryFrom<(Ident, Group)> for GodotConfigurationPredicate {
 
     fn try_from((ident, group): (Ident, Group)) -> Result<Self, Self::Error> {
         match ident.to_string().as_str() {
-            "any" => Ok(Self::Any(Box::new(GodotConfigurationAny::try_from(group)?))),
-            "all" => Ok(Self::All(Box::new(GodotConfigurationAll::try_from(group)?))),
-            "not" => Ok(Self::Not(Box::new(GodotConfigurationNot::try_from(group)?))),
+            "any" => Ok(GodotConfigurationAny::try_from(group)?.into()),
+            "all" => Ok(GodotConfigurationAll::try_from(group)?.into()),
+            "not" => Ok(GodotConfigurationNot::try_from(group)?.into()),
             _ => Err(Self::Error::UnableToParse(format!(
                 "unrecognised ident: {}",
                 ident
@@ -105,255 +142,11 @@ impl TryFrom<(TokenTree, TokenTree)> for GodotConfigurationPredicate {
     }
 }
 
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct GodotConfigurationAll(GodotConfigurationPredicateList);
-
-impl TryFrom<Group> for GodotConfigurationAll {
-    type Error = GodotConditionalCompilationError;
-
-    fn try_from(group: Group) -> Result<Self, Self::Error> {
-        Ok(Self(GodotConfigurationPredicateList::try_from(group)?))
-    }
-}
-
-impl GodotConditionalCompilation for GodotConfigurationAll {
-    fn should_compile(&self) -> bool {
-        self.0.iter().all(|option| option.should_compile())
-    }
-}
-
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct GodotConfigurationAny(GodotConfigurationPredicateList);
-
-impl TryFrom<Group> for GodotConfigurationAny {
-    type Error = GodotConditionalCompilationError;
-
-    fn try_from(group: Group) -> Result<Self, Self::Error> {
-        Ok(Self(GodotConfigurationPredicateList::try_from(group)?))
-    }
-}
-
-impl GodotConditionalCompilation for GodotConfigurationAny {
-    fn should_compile(&self) -> bool {
-        self.0.iter().any(|option| option.should_compile())
-    }
-}
-
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct GodotConfigurationNot(GodotConfigurationPredicate);
-
-impl TryFrom<Group> for GodotConfigurationNot {
-    type Error = GodotConditionalCompilationError;
-
-    fn try_from(group: Group) -> Result<Self, Self::Error> {
-        let tokens: Vec<_> = group.stream().into_iter().collect();
-        match tokens.len() {
-            1 => {
-                // ToDo: Make this more efficient
-                let mut iter = tokens.into_iter();
-                let likely_ident = iter.next().unwrap();
-                Ok(Self(GodotConfigurationPredicate::try_from(likely_ident)?))
-            }
-            2 => {
-                // ToDo: Make this more efficient
-                let mut iter = tokens.into_iter();
-                let likely_ident = iter.next().unwrap();
-                let likely_group = iter.next().unwrap();
-                Ok(Self(GodotConfigurationPredicate::try_from((
-                    likely_ident,
-                    likely_group,
-                ))?))
-            }
-            n => Err(Self::Error::UnableToParse(format!(
-                "not may only have one predicate inside it, found {}: {}",
-                n, group
-            ))),
-        }
-    }
-}
-
-impl GodotConditionalCompilation for GodotConfigurationNot {
-    fn should_compile(&self) -> bool {
-        !self.0.should_compile()
-    }
-}
-
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct GodotConfigurationPredicateList(Vec<GodotConfigurationPredicate>);
-
-impl Deref for GodotConfigurationPredicateList {
-    type Target = Vec<GodotConfigurationPredicate>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl TryFrom<Group> for GodotConfigurationPredicateList {
-    type Error = GodotConditionalCompilationError;
-
-    fn try_from(group: Group) -> Result<Self, Self::Error> {
-        let mut inner = vec![];
-        let mut iter = group.clone().stream().into_iter().peekable();
-        loop {
-            let next = iter.next();
-            if next.is_none() {
-                break;
-            }
-            let first_token = next.unwrap();
-
-            match iter.next() {
-                // Single word
-                None | Some(TokenTree::Punct(_)) => {
-                    inner.push(GodotConfigurationPredicate::try_from(first_token)?)
-                }
-                // Word followed by group
-                Some(TokenTree::Group(group)) => {
-                    match iter.peek() {
-                        // Group is last token or followed by punctuation
-                        None | Some(TokenTree::Punct(_)) => {
-                            let _ = iter.next(); // Skip the punctuation
-                            inner.push(GodotConfigurationPredicate::try_from((
-                                first_token,
-                                group.into(),
-                            ))?)
-                        }
-                        // Anything else
-                        _ => return Err(Self::Error::UnableToParse(group.to_string())),
-                    }
-                }
-                // Anything else
-                _ => return Err(Self::Error::UnableToParse(group.to_string())),
-            }
-        }
-        Ok(Self(inner))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use proc_macro2::TokenStream;
     use std::str::FromStr;
-
-
-
-    #[test]
-    fn test_predicate_list() {
-        let ts = TokenStream::from_str("(test, doctest)").unwrap();
-        let tt = ts.into_iter().next().unwrap();
-        if let TokenTree::Group(group) = tt {
-            let list = GodotConfigurationPredicateList::try_from(group).unwrap();
-            assert_eq!(
-                list,
-                GodotConfigurationPredicateList(vec![
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption ::Test),
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption ::DocTest),
-                ])
-            )
-        } else {
-            panic!("Token tree was not a group")
-        }
-    }
-
-    #[test]
-    fn test_all() {
-        let mut ts = TokenStream::from_str("all(test, test)")
-            .unwrap()
-            .into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::All(Box::new(GodotConfigurationAll(
-                GodotConfigurationPredicateList(vec![
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption::Test),
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption::Test),
-                ])
-            )))
-        );
-
-        let mut ts = TokenStream::from_str("all(test, doctest)")
-            .unwrap()
-            .into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(!predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::All(Box::new(GodotConfigurationAll(
-                GodotConfigurationPredicateList(vec![
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption::Test),
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption ::DocTest),
-                ])
-            )))
-        );
-    }
-
-    #[test]
-    fn test_any() {
-        let mut ts = TokenStream::from_str("any(test, doctest)")
-            .unwrap()
-            .into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::Any(Box::new(GodotConfigurationAny(
-                GodotConfigurationPredicateList(vec![
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption ::Test),
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption::DocTest),
-                ])
-            )))
-        );
-
-        let mut ts = TokenStream::from_str("any(doctest)").unwrap().into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(!predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::Any(Box::new(GodotConfigurationAny(
-                GodotConfigurationPredicateList(vec![GodotConfigurationPredicate::Option(
-                    GodotConfigurationOption::DocTest),])
-            )))
-        );
-    }
-
-    #[test]
-    fn test_not() {
-        let mut ts = TokenStream::from_str("not(test)").unwrap().into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(!predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::Not(Box::new(GodotConfigurationNot(
-                GodotConfigurationPredicate::Option(GodotConfigurationOption ::Test)
-            )))
-        );
-
-        let mut ts = TokenStream::from_str("not(not(test))").unwrap().into_iter();
-        let ident = ts.next().unwrap();
-        let group = ts.next().unwrap();
-        let predicate = GodotConfigurationPredicate::try_from((ident, group)).unwrap();
-        assert!(predicate.should_compile());
-        assert_eq!(
-            predicate,
-            GodotConfigurationPredicate::Not(Box::new(GodotConfigurationNot(
-                GodotConfigurationPredicate::Not(Box::new(GodotConfigurationNot(
-                    GodotConfigurationPredicate::Option(GodotConfigurationOption::Test)
-                )))
-            )))
-        );
-    }
 
     #[test]
     fn test_everything() {
@@ -372,7 +165,9 @@ mod tests {
                         GodotConfigurationPredicateList(vec![
                             GodotConfigurationPredicate::Option(GodotConfigurationOption::Test),
                             GodotConfigurationPredicate::Not(Box::new(GodotConfigurationNot(
-                                GodotConfigurationPredicate::Option(GodotConfigurationOption::DocTest)
+                                GodotConfigurationPredicate::Option(
+                                    GodotConfigurationOption::DocTest
+                                )
                             ))),
                         ])
                     ))),
